requirements.txt
Flask
whisper
sounddevice
numpy
pyttsx3
scipy
python-dotenv
groq
